+++ src/.internal/core/render/Rectangle.ts
@@ -1,6 +1,11 @@
 import { Graphics, IGraphicsSettings, IGraphicsPrivate } from "./Graphics";
 
 export interface IRectangleSettings extends IGraphicsSettings {
+	/**
+	 * @todo review
+	 * If this is set to `true`, rectangle will be drawn in such a way that its stroke is contained within the rectangle's width and height.
+	 */
+	containStroke?: boolean;
 }
 
 export interface IRectanglePrivate extends IGraphicsPrivate {
@@ -20,7 +25,7 @@ export class Rectangle extends Graphics {
 	public static className: string = "Rectangle";
 	public static classNames: Array<string> = Graphics.classNames.concat([Rectangle.className]);
 
-	public _afterNew(){
+	public _afterNew() {
 		super._afterNew();
 		this._display.isMeasured = true;
 		this.setPrivateRaw("trustBounds", true);
@@ -43,7 +48,23 @@ export class Rectangle extends Graphics {
 	}
 
 	protected _draw() {
-		this._display.drawRect(0, 0, this.width(), this.height());
+		let w = this.width();
+		let h = this.height();
+		let x = 0;
+		let y = 0;
+
+		let wSign = w / Math.abs(w);
+		let hSign = h / Math.abs(h);
+
+		if (this.get("containStroke", false)) {
+			const strokeWidth = this.get("strokeWidth", 0);
+			w -= strokeWidth * wSign;
+			h -= strokeWidth * hSign;
+			x += strokeWidth / 2 * wSign;
+			y += strokeWidth / 2 * hSign;
+		}
+
+		this._display.drawRect(x, y, w, h);
 	}
 
 	public _updateSize() {