+++ src/stock.ts
@@ -34,6 +34,7 @@ export { VolumeProfile, IVolumeProfileEvents, IVolumeProfilePrivate, IVolumeProf
 export { CommodityChannelIndex, ICommodityChannelIndexEvents, ICommodityChannelIndexPrivate, ICommodityChannelIndexSettings } from "./.internal/charts/stock/indicators/CommodityChannelIndex";
 export { DisparityIndex, IDisparityIndexEvents, IDisparityIndexPrivate, IDisparityIndexSettings } from "./.internal/charts/stock/indicators/DisparityIndex";
 export { StandardDeviation, IStandardDeviationEvents, IStandardDeviationPrivate, IStandardDeviationSettings } from "./.internal/charts/stock/indicators/StandardDeviation";
+export { SuperTrend, ISuperTrendEvents, ISuperTrendPrivate, ISuperTrendSettings } from "./.internal/charts/stock/indicators/SuperTrend";
 export { TypicalPrice, ITypicalPriceEvents, ITypicalPricePrivate, ITypicalPriceSettings } from "./.internal/charts/stock/indicators/TypicalPrice";
 export { Trix, ITrixEvents, ITrixPrivate, ITrixSettings } from "./.internal/charts/stock/indicators/Trix";
 export { MedianPrice, IMedianPriceEvents, IMedianPricePrivate, IMedianPriceSettings } from "./.internal/charts/stock/indicators/MedianPrice";